Player stats + slow-download UX, wired torrent profiles, HDR fix (v0.1.16)
- HDR: target-colorspace-hint=no so gpu-next tone-maps HDR to SDR in the
  composited WebView2 window (was rendering DV/HDR10 washed grey).
- Buffering screen: rename to "Initializing"; escalate after sustained
  near-zero download to "Try a different source" / "Switch to Fast mode"
  (honest VPN/IP caveat), plus a Cloudflare speed test to tell a slow
  source from a slow connection.
- Streaming server: wire the previously-vestigial torrent profiles;
  download hard-limit maps to librqbit ratelimits (live), Ultra Fast is
  now the default, "Default" profile removed.
- Player nav: remove the redundant header fullscreen button (fullscreen
  stays in the window chrome and control bar).
- Statistics menu: enable the button whenever a stream is active (was
  greyed out for single-file torrents and direct streams); drop torrent
  stats for non-torrent streams; stop asserting "SDR"/"Software" before
  mpv reports real video-params (show a placeholder until known).
- Bump version to 0.1.16.

+++ crates/streaming-server/src/types.rs
@@ -49,8 +49,10 @@ pub struct SettingsValues {
     pub bt_max_connections: u64,
     pub bt_handshake_timeout: u64,
     pub bt_request_timeout: u64,
-    pub bt_download_speed_soft_limit: u64,
-    pub bt_download_speed_hard_limit: u64,
+    // f64 (not u64): core reads these as f64 and the "soft" profile uses a
+    // fractional B/s value, so u64 truncation would make the UI show "custom".
+    pub bt_download_speed_soft_limit: f64,
+    pub bt_download_speed_hard_limit: f64,
     pub bt_min_peers_for_stable: u64,
     pub remote_https: String,
     pub local_addon_enabled: bool,
